Implement pre migration and post migration endpoints
Since the virt-launcher will stop itself if no VM appears or when a VM
disappears, the virt-launcher pod retry policy was changed to Never.
This guarantees that after a migration the old pod goes down and will
not be restarted.

With the current POC state we don't have any chance to avoid race
conditions. Especially since we can't change pod names as a resource
controller outside of kubernetes. Therefore we can't rely on the
uniquenes of our VM pod. Further we don't have our own data store yet to
guarantee any uniquenes. There are a lot of possibilities that we und up
with the same VM started multiple times. Only a racy check tries to
prevent the worst.

Especially since the initial virt-launcher pod is responsible for
starting a VM this all does not really fit for migrations now.

Change-Id: I89791afa8dfa93b68fdd1c275e327bb9fb8d328c
Signed-off-by: Roman Mohr <rmohr@redhat.com>