Check if go files pass "fmt" and "vet" on travis
The "fmt" check is a little bit hacky since it checks if "go fmt"
returns a file to format. Since this involves a normal format run, after
that the file is formated. It would be cleaner to wrap "gofmt -l" in a
script and run it on all scripts to not modify wrong formatted files
while checking.