controller: Drop explicit docker registry flag
Currently the virt-controller provides a --docker-registry flag, this can be
used to specify a registry to be used for images used by the controller, in
fact we are currently just using it for the launcher image.

But we actually do not need to separate the registry from the image name,
because both can be specified in one string (<registry>/<ns>/<image>:<tag>).

By dropping the flag and switching to the scheme above, the reference of
images is now aligned in all places in kubevirt.
Now all images are referenced by this scheme.

Signed-off-by: Fabian Deutsch <fabiand@redhat.com>