Remove useless Test for Jobs
This test was supposed to ensure that not remote calls were
made if the labels didn't match.  However, the matching of labels
was not done by code executed in this test, and cannot be done with
out a live server.  Thus, the test should have been failing, and
only passed due to a second error that hid the fact that call had
been made.