Add simple, consistent libvirt reconnect strategy
Make sure, that whenever an error happens which indicates that we lost
the connection to libvirt, that we mark the connection as broken and
immediately return the error instead of reconnecting.

On the next call, try to repair the function before doing the actual
call. This repair attempt will also trigger re-registration of event
listeners, which will then resynchronize with libvirt.

Add a watchdog, which checks every 10 seconds, if the connection
is alive, and marks the connection as broken if not.

Finally this fixes the two functional tests, which were waiting to
receive crash events from a VM.