Ensure libvirt client auth config file is always used
If you call a plain libvirt.NewConnect, libvirt will never consult
its client side auth config file. So even if the username & password
we have are both "", we should still call NewConnectWithAuth to
enable this config file lookup. This lets the user create a auth.conf
containing something like

  [credentials-test]
  authname=test
  password=123456

  [auth-libvirt-localhost]
  credentials=test

and that'll be automatically used when virt-handler/virt-manager
connect to libvirtd, if auth is requested by libvirtd.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>