Introduce a "graphics" field on the VM object status
Record the VNC/SPICE graphics port details against the
VM object, under the VMStatus struct. This gets populated
with the (potentially dynamically allocated) QEMU port
for VNC/SPICE, and the IP address seen from virt-handler,
which is assumed to be the same as the IP seen from
libvirtd, given both are in the host net namespace.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>