vagrant: Work around controller race bug until Kubernetes 1.7.1 is released
When nodes try to join a master, they can fail because cluster-info is
not updated with the expected tokens. To work around that, add the
required Role and RoleBinging to let the token signer do its work in
time. See https://github.com/kubernetes/kubeadm/issues/335 for details
about the workaround.

Signed-off-by: Roman Mohr <rmohr@redhat.com>