docs: add instructions for a non-vagrant devel environment
Not everyone will want to use vagrant for the development
environment. Particularly if their primary machine is
already a VM, it makes no sense to spin up further VMs
since they likely can't use KVM nested. It is quite simple
to run an all-in-one local k8s environment with the basic
commands.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>