tests: Wait until console handler is done before checking mocks
In the virt-handler unit tests, the httptest server sometimes reports
already that the server is done, before the handler method call is done.
This leads to flaky checks on the expected mocked out calls. Therefore,
make sure that the http handler is done, by explicitly waiting for a
channel, which signals that.

This fixes temporary test errors like

  missing call(s) to *virtwrap.MockVirDomain.Free()

Signed-off-by: Roman Mohr <rmohr@redhat.com>