Create basic RBAC permissions for KubeVirt
Make sure that KubeVirt can run on Kubernetes instances where RBAC is
enabled. Further give the haproxy admin rights right now, to allow it to
forward all calls to the apiserver.

This does not at all provide a more secure way to run KubeVirt. It's
only purpose at this stage, is to allow friction less deployments on
RBAC enabled Kubernetes clusters.

The depolyment can be properly secured, as soon as we have movedm from
our haproxy approach to the Aggregated API Server.

Signed-off-by: Roman Mohr <rmohr@redhat.com>