Don't check for exact IPs in storage tests
In many cases, we don't see the Pod IP connecting. Therefore, remove all
workarounds, to cope with these scenarios, and only check if a new
connection appears when the VM is started. This simplifies the tests and
is almost as good, since the BeforeEach block makes sure that no
connection exists before each test.

Signed-off-by: Roman Mohr <rmohr@redhat.com>