Changes ContainerRegistryDisk readiness checks
Previously we were using a TCP probe as a readiness check. This
involved a race condition where the iscsi server could be online but
the LUN was not yet registered.

Now the registry disk base container touches a file when the
setup process is complete. The readiness check looks for this
file instead of doing the tcp check.

Signed-off-by: David Vossel <dvossel@redhat.com>