document v1.VM and api.DomainSpec
Both v1 and api contain structures that are essential to kubevirt. It
is easy to confuse them as right now, v1.VM contains a structure
identical to api.DomainSpec. To help with the confusion, we add a
godoc compatible to the structures, making it easy to distinguish
which structure we are dealing with.

Signed-off-by: Martin Polednik <mpolednik@redhat.com>