Improve VirtualMachineReplicaSet tests
Add a mock workqueue, which allows waiting for an expected number
enqueues. This allows synchronous testing of the controller. The typical
pattern is:

  mockQueue.ExpectAdd(3)
  vmSource.Add(vm)
  vmSource.Add(vm1)
  vmSource.Add(vm2)
  mockQueue.Wait()

This ensoures that informer callbacks which are listening on vmSource
enqueued three times an object. Since enqueing is typically the last
action in listener callbacks, we can assume that the wanted scenario for
a controller is set up, and an execution will process this scenario.

Signed-off-by: Roman Mohr <rmohr@redhat.com>