Make it observable if the VMRS is in paused state
Once the VMRS gets paused, it adds a Paused condition to it's state, so
that it is possible to observe from outside if the controller already
stopped. Only then is it save to modify/delete VirtualMachines which the
controller controlls.

That allows cluster-autoscaler to wait until the controller stopped,
then delete unused nodes and finally enable the controller again with a
decremented replica count.

Signed-off-by: Roman Mohr <rmohr@redhat.com>