set standard for flag handling in virt-handler
The current flag handling has few issues:

* it is mixed with initiation code,
* several flags are duplicated throughout services,
* there is no common interface or guidelines for flag handling.

We solve this by improving the flag handling (for start) in
virt-handler. Common flags are now handled by service package
(-listen, -port), further flags can be added within the service. This
is done in AddFlags() function of (to be) Service interface. The
interface is not yet declared as we now have compile-time interface
conformance checking, therefore all services need to use this flag
handling before it can be declared.

Signed-off-by: Martin Polednik <mpolednik@redhat.com>