ovm: support for stable firmware UUID
When a VM starts from one OfflineVM, it should
always have the same firmware UID. We call this fact "stable UUID".

Previously, unless the user specified the UUID in the OfflineVM,
each time the corresponding VM was started, a random firmware
was used.

This can confuse some operating systems (Windows being a prime example),
so we add the possibility to fix the UID.
The simplest approach, implemented in this patch, is to use a V5 UUID
using the VM name and a custom namespace UUID as source values.

Signed-off-by: Francesco Romani <fromani@redhat.com>