kubecli: introduce StreamInterface
Previously, because of how the API looked like, we had no way
to get an handle to a VM and enable the data flow for VNC/console
in two separate steps: everything was done at once:

Establish a connection to the serial console AND move the data around:

  err := virtCli.VM(namespace).SerialConsole(vm, stdinReader, stdoutWriter)

That worked fine, but if the client code needed to do any preliminar
validation (e.g. checking if it could reach the VM), it had either to
re-implement the check, or to blindly call the API and hope for the
best.

With this patch, we change the API to separate the connection and
the data management:

  con, err := virtCli.VM(namespace).SerialConsole(vm)
  if err == nil {
    err = con.Stream(kubecli.StreamOptions{
      In:  stdinReader,
      Out: stdoutWriter,
    })
  }

This change makes the API more flexible and not less powerful, and
enable any client code to do connection or post-connection checks.

Reuses ideas from https://github.com/kubevirt/kubevirt/pull/1015