Fix oc template samples
The template samples within the 'cluster' directory use double curly
braces for CPU_CORES and MEMORY variables and single curly braces for
NAME variable.  Since user is not expected to enter Json string as a
value for CPU_CORES and MEMORY parameters, this is quite confusing,
especially for the new KubeVirt users.