Add clusterrole aggregation for kubevirt CRDs
By default regular users do not have access admin/edit/view access to
CRDs such as those provided by kubevirt.  This causes problems when a
user wants to view these objects in a UI or provision them.  By adding
aggregation labels to the clusterroles the user's existing permissions
for regular objects will be mapped to our CRDs.  This provides a better
experience for users.  Admins who want to lock down kubevirt access
should be able to add it on later.

Signed-off-by: Adam Litke <alitke@redhat.com>