Propagate API macAddress to bridge backend
Ultimately we override mac_address in libvirt interface configuration, which
makes the guest to see the overridden MAC address.

This patch adds functional test to validate that the overridden MAC address is
indeed seen inside the guest OS; and that instances with overridden MAC
addresses can communicate with each other and the outside world.

Note that this patch does *not* provide support for custom MAC addresses for
slirp interfaces. This will be followed up on in next patches in the series.