Add a single test to verify connectivity for explicit given pod network
Part 1

Don't run all network tests twice, once with explicit pod network
definition given and once with the implicit assumption.

Instead of that create a third VM which has an explicit network
definition given and add add an explicit test which ensures that it is
connected to the pod network.

This also fixes flakyness in the tests, since pod and service definition
conflicts occured from time to time, or the service tried to reach a vm
which was not intended by the test.