Introduce an optional subprotocol for browser websocket connections
Browsers need to use subprotocols to pass credentials to kubernetes to
open websocket connections. Kubernetes will remove the authentication
subprotocol from the protocols. In order to let the browser continue the
handshake, the server needs to return another subprotocol. The
subprotocol is named "plain.kubevirt.io" and is an alias for a plain
websocket connection.

A valid connection attempt from the browser may look like this:

  Sec-WebSocket-Protocol: base64url.bearer.authorization.k8s.io.bXl0b2tlbg, plain.kubevirt.io

The server will then respond with

  Sec-WebSocket-Protocol: plain.kubevirt.io

If no subprotocols are given, the connection is upgraded regardless, for
backward compatibility.