Don't use tty in remote exec
virtctl vnc didn't work with CRI-O, because the remote exec command in
virt-api converted LFs into CR + LF, which broke the VNC network protocol.
The root cause for this is that this seems to be a default behaviour
of tty, see https://github.com/moby/moby/issues/8513#issuecomment-216191236.
The solution is to not use tty, we don't need it.
(Still unclear why this only happened with CRI-O)

Signed-off-by: Marc Sluiter <msluiter@redhat.com>