Set option 15 for windoes machines on DHCP requests
In order to allow service-name DNS resolution not only via the FQDN, it
is necessary to put the VMI into a domain. This domain equals to the
search domain entry of the namespace the VMI runs in.

As a consequence, services in the namespace of the VMI can be resolved

 * via the FQDN: myservice.mynamespace.svc.cluster.local
 * relative the the namespaces: myservice

Before only FQDN was possible.