Fix test to pass within STDCI environment
It seems the 'latest' tag is not valid within STDCI environment anymore
for Fedora image--we should use 'devel' tag instead.  The 'devel'
happens to be the default Docker tag in the example Fedora template, so
we can directly call GetTemplateFedora() function from the
vms-generator.

The GetTestTemplateFedora() function, introduced with the earlier
versions of the Template test, became obsolete and it was deprecated.