Fix support for multiple SR-IOV networks referring same resource
If there are multiple SR-IOV networks attached to a VMI, and some of them refer
to the same resource name in their annotations, SR-IOV device plugin receives
request for >1 device of the same name. In this case, it passes multiple device
IDs using the same variable.

In DP v1.0, these IDs were split with commas. But in the latest master, they
are now split with space characters. This patch adopts to the new format.

Co-Authored-By: Greg Bock <greg.bock@stackpath.com>