Increase test timeouts on many places
We unintentionally waited twice as long for a VMI start than we claimed
to in one of our test helper functions, due to the fact that we did not
check if the start event occured. If it did not occure we just polled
again for the same amount of time afterwards.

Further give pods which query VMIs over network more time to bescheduled
and to complete.

Adjust more timeouts