Only add infra pod when a readiness probe is specified
Starting the pod takes slightly longer. On stdci it has a huge impact on
the test runtime. Also we don't want the starts to take longer than
necessary. Since we will need the infra pod to be always there in order
to have a better default readiness probe for the VMI, we will need to
find a solution.

Two possibilities:

 * Improve the libvirt startup time (more than 10 seconds, even on
faster systems).
 * Reduce the impact on the pod start time when an additional container
is added.