virtwrap: add domainstats support
We add support for getting the domain stats inside the virtwrap package.
To do so, we wrap the libvirt Bulk Stats APIs (the preferred and
recommended way to accomplish this task) in the virtwrap subpackage.

However, we chose not to expose the libvirt data types directly, thus
we add wrappers and quasi-clones of the data types in the stats/
subpackage, together with the converter.

For the sake of the practicality, and to avoid dependency cycles
between packages, we augment the cli subpackage with a simple facade
the client code may use to fetch all the relevant stats.

See the package docs for more details

Signed-off-by: Francesco Romani <fromani@redhat.com>