metrics: gather metrics in separate goroutines
We use the libvirt stats API to gather the domain stats through
the socket protocol.
The problem is these APIs may block for indefinite amount of time,
hours or even forever, should QEMU get stuck in I/O.

This is a well known problem observed in other management solutions
(like oVirt) or metric-gathering systems (like collectd) and, due
to some limitations of the QEMU management protocol, will not go
away anytime soon.

In the worst case, a call to GetDomainStats may block forever.
We need to deal with two possible issues arising from this:
1. we need to be able to send management commands to the VM even if
   a call is stuck gathering metrics.
2. we need to minimize the impact of a rogue misbehaving VM.
   A call being stuck should not harm the collection of the metrics
   of the other, well behaving VMs.

For point #1, it seems that the net/rpc go package we use to implement
our socket protocol already gives us enough safety:
Quoting https://golang.org/pkg/net/rpc/#Client:
"""
Client represents an RPC Client. There may be multiple outstanding Calls
associated with a single Client, and a Client may be used by multiple
goroutines simultaneously.
"""

It is then possible to send other commands to the socket while a
GetDomainStats call is ongoing.

For point #2, addressed by this patch proper, we just gather the metrics
for the VMs each own in a separate throwaway goroutine, which will end
once the update is done.

Signed-off-by: Francesco Romani <fromani@redhat.com>