What this PR does / why we need it: Changed all test levels in api_validation_test to focus (FDescribe, FContext, FIt) added rfe id, test id, criticality etc
based on issue
https://github.com/kubevirt/containerized-data-importer/issues/359

Special notes for your reviewer:
This is an example on how i plan to change the tests, so we will be able to auto create them in Polarion from the code.
There are 2 steps to it:
first, we create the tests without test ids (if they are new tests)
after running the script which creates the tests in Polarion (kubevirt/qe-tools/polarion-generator), each one gets its own ID,
and we can than add the test ids to the test code (this will allow reporting & changing the test in the future).