Move SR-IOV functional tests into a separate suite
This will allow us to target just SR-IOV specific tests for SR-IOV test
lane. This also makes it so that these tests are not even triggered for
multus CI jobs since we know we don't deploy SR-IOV components there.

In the future we will add a CI job for SR-IOV that will run the new
SRIOV suite.