collector: fix the concurrentCollector lifecycle
The lifecycle of the concurrentCollector instance must be tied to
the one of the Collector instance. Previously we had one instance per
promethues.Collect(), which works but defeats completely the purpose
of tracking the busy sources.

The tests were working since they assumed a concurrentCollector instance
was preserved across prometheus.Collect() calls, which is indeed the right
thing to do, so this patch fixes this.

Signed-off-by: Francesco Romani <fromani@redhat.com>