tests: retry scraping before give up
We added a test to scrape the prometheus endpoint looking for VM
metrics.

Most of the time the tests fails, it is because it fails to connect to the
cmd-client socket.
Not sure why, because the test waits for the VMI to be Running before scraping
(thus attempting to connect). The most likely reason seems to be
slow/overloaded CI.

To improve the test, we wrap the scraping call in an Eventually(),
retrying few times before to give up.

Signed-off-by: Francesco Romani <fromani@redhat.com>