Fix swagger spec generation for our subresources
We only generate the openapi spec once and then serve the content from
memory for the subresources. By accident the variable which is used as
storage was also part of the once closure, which lead to only serving
the spec once and null in subsequent calls.

The symptome was an error in the apiserver log:

  E0207 20:04:08.320152       1 controller.go:165] updating
    "v1alpha3.subresources.kubevirt.io" to AggregationController failed
    with: ERROR $root is missing required properties: info, swagger