hyperv: add feature gate and node selector
Some HyperV enlightenments need an up to date host kernel to run
properly. Should the kernel not support a HyperV enlightenment,
QEMU won't start and the VM won't run.

Unfortunately QEMU can't detect and report the KVM support status
in the kernel ahead of time: all the checks are performed in
the VM startup flow, when it is too late to recover.

To overcome this limitation, we require host labels to advertise
the support of the required HyperV enlightenments.
We will add nodeSelectors to make sure HyperV-enabled VMs are
scheduled only on nodes which support all the required enlightenments.

How the labels gets added doesn't concern KubeVirt itself:
KubeVirt just assumes the label are correctly added by some external
entity.
A NFD addon is being prepared to do the very same checks QEMU does
to label the nodes.

To be backward compatible, the additional nodeSelectors are disabled
by default. This depends on a feature gate.

It will work like this:
- feature gate DISABLED (default for backward compat)
  hyperv-enabled VMs will run in best-effort on any node in the
  cluster.
- feature gate ENABLED:
  hyperv-enabled VMs will run only on nodes which have the right
  set of hyperv labels.

Finally, we must point out that in the common case, up to date
host should support all the hyperv enlightenmets KubeVirt supports:
to lack a feature (say, SyNIC) is deemed rare, although possible.

Signed-off-by: Francesco Romani <fromani@redhat.com>