Fix numbering for multus interface names
We should not count interfaces that are not handled by multus, otherwise
we may request an interface with incorrect name from multus, which then
makes the corresponding VMI to fail to start because of missing netlink
device with the name (e.g. multus would then create an interface called
`net2` while we would expect `net1`).