Add functests to check that custom MAC address not in pod netns
If we pass MAC address to CNI plugin, there is a risk that the address
will be set inside the pod namespace, in addition to the VM guest, which
may disrupt connectivity. KubeVirt should make sure that the same MAC
address is never used both inside the guest and inside pod namespace.

The new tests cover bridge interface mode only; slirp and masquerade
modes are not convered because they support pod networks only.