metrics: add test for prometheus throttling
Add a test to make sure the throttling is active.
Unfortunately, it is quite hard to have reliable failures without
injecting delays in the code.
The only good enough way emerged so far is to use a large (> 200)
amount of clients with very short timeouts, but this still has
a chance of false negatives.

Signed-off-by: Francesco Romani <fromani@redhat.com>