Fixed grpc dependency of device manager
This took me hours: bazel files of device manager are manually maintained,
and they use a different way of declaring the grpc dependency than the
generated bazel files. For some reason that resulted in initializing
Go's trace package twice... which failed, because it can't register the
same debug path to the server twice.

Signed-off-by: Marc Sluiter <msluiter@redhat.com>