Create the ovs br1 on worker nodes for ocp 3.11
This PR fixes an issue on the ocp provider.

When we request more the one node for the cluster the additional workers
are reinstalled using the openshift-ansible scale-up playbook.

This playbook removes all the ovs configuration done previously,
so this PR just run on every node using the ovs daemonset installed in the
cluster and recreate the br1 bridge if not exist.