pkg: lookup: scan VMIs on a node
In order to augment the metrics labels with VMI info
(see: https://github.com/kubevirt/kubevirt/issues/2108), the
prometheus collector needs to know the VMIs which are running
on each node.

Up until now, the collector just scanned the socket directory,
inferring everything it needed for VMs (in libvirt sense) from
just the data got from the socket channel. It used to work like this
1. get the list of all the cmdclient sockets in the host
2. query each (libvirt) VM for stats connecting to each socket
3. build metrics based on the stats

This is not enough anymore, so the collection process changes as follow:
1. get the list of all the *VMIs* in the host
2. infer the cmdclient socket path from the VMI spec (there are
supported APIs for that, so this step is safe)
3. query each (libvirt) VM for stats connecting to each socket
4. build metrics based on the stats and the VMI spec

This patch implements the refactoring, extracting the function we
need to get the list of VMIs running from the virt-controller,
and moving in a new util pkg.

No changes are yet exposed to the metrics layer, just refactoring
and preparation

Signed-off-by: Francesco Romani <fromani@redhat.com>