Use sriov network multus name from HCO for VMI example
While there is no canonical name for the network, it seems like a
better choice to use the name of the network that is produced by
hyperconverged operator (HCO) that is also maintained by kubevirt
org. This allows to apply the example VMI manifest to a cluster deployed
with HCO without any changes. (Except perhaps the image registry.)