hyperv: enable dep check for requested features
Some hyperv features depend on each other
(see https://schd.ws/hosted_files/devconfcz2019/cf/vkuznets_enlightening_kvm_devconf2019.pdf for full reference),
yet neither libvirt nor qemu enforce this dependency chain.
For example, vpindex is required for synic, ipi. But if the user is not aware, or forgets to enable it and yet
it requires synic, the VM won't boot.

This patch implements a dependency check. We examine the dependencies
between hyperv features, rejecting the VM ahead with a clear error
message if needed.

The alternative is do nothing and let QEMU ultimately reject the VM.
I believe this is worse UX, because in this case the error is buried
deeply in the QEMU/virt-launcher logs, and the user has to dig it out.

Resolves: https://github.com/kubevirt/kubevirt/issues/2172
Signed-off-by: Francesco Romani <fromani@redhat.com>