Make secret lookup of virt-handler and virt-api more robust
virt-handler and virt-api create certificates and upload them in a
secret if it does not already exists. When deploying or updating
kubevirt the instances tend to do the GET and POST at roughly the same
time which leads to restarts of the instances which loose the race.

Handle this gracefully by always first try to to a POST and fall back if
the secret already exists. This avoids unnecessary restarts during
initialization.