libvirt: expose a way to enable the  debug logs
In order to make easier to troubleshoot VM launch failure,
this patch adds a way to enable the libvirt debug logs.
See https://wiki.libvirt.org/page/DebugLogs for details.

In order to set the expectactions straight, libvirt logs
are often tremendously helpful to understand what the VM is doing,
but are not always enough; furthermore, QEMU logs are not affected
by this change, and will keep being very terse (note this has to
be fixed with QEMU changes).

To enable the libvirt logs, you need to add the `debugLogs=true`
annotation to the VMI spec. This is completely optional; by
default the libvirt logs are kept the way are now.

Signed-off-by: Francesco Romani <fromani@redhat.com>