Increase timeout on PDB test for eviction strategies
How long it takes until a deleted pod, which is scheduled but not quite
running yet, is very unpredictable. Even if there is no load on the node
it can vary greatly and can depend on kubelet internals of different k8s
versions.

Increasing the timeout for one of the tests where we have to waid for
the VMI and its pod to disappear.

Signed-off-by: Roman Mohr <rmohr@redhat.com>