Expand sriov docs with brief architecture notes
KubeVirt documentation is not the right place to expand too much on how
all components fit, but the brief notes should hopefully be a good start
for an interested reader to dive in if needed.

Also, be more explicit SR-IOV should be deployed with operator.

Plus, restored some documentation that was unintentionally removed in
a previous patch that still belongs in the document as part of
description of manual deployment of relevant SR-IOV components if an
admin / developer decides to not follow the operator path.

Finally, added some external links to relevant docs / blogs.

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>