Allow as many scrapes on the VMI, like we allow parallel prom scrapes
Tie the allowed number of metric scrapes on individual qemu processes to
the amount of parallel prometheus scrapes on the API.

This avoids errors when more than one scrape goes on in parallel, and a
VMI is currently being scraped by more than one call. In this case, no
metrics were collected, which lead to missing metrics in prometheus.

Signed-off-by: Roman Mohr <rmohr@redhat.com>