Move containerdisks to their own BUILD file
Avoid pulling in all testing and demo container dependencies if one just
wants to build kubevirt and run unittests.

Before we had the containerdisk definitions at the same location like
the containerdisk binary and  we were just pulling them in on every
"make build".

Now the containerdisks are only pulled in when we explicitliy want to
build and/or push containers.

Signed-off-by: Roman Mohr <rmohr@redhat.com>