tests: give stress tool time to trash some memory pages before migration
Right now we immediately start migration after stress test tool is started. If
migration process is quite quick, it may overtake the stress tool and migrate a
VMI successfully when a failure would be expected otherwise.

This scenario happened more prominently with changes that moved network
configuration from launcher to handler because of changes in timing of
components (apparently, migration may start earlier than before); though it's
still possible to produce a spurious failure of the tests even on clean tree.

This patch gives 15 seconds of time for each stress tool start to make it less
likely the migration succeeds when a failure is expected.

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>