Ensure our metrics labels are not overwritten
Since OCP 4.3, the monitoring stack is featuring a Prometheus proxy (Thanos).
One of the proxy components (prom-query-proxy) is modifying the metrics on the
fly: it changes our namespace label to exported_namespace. In result, while
virt-handler sets the namespace label to be the namespace in which the VMI
resides, it's being overwritten to the namespace of virt-handler.
Since virt-handler's namespace is privileged, users without access to this
namespace will not be able to view this metric.

Signed-off-by: Daniel Belenky <dbelenky@redhat.com>