Drop support for Genie networks
Our Genie test lanes are utterly updated and there does not seem to be users of
this binding anyway. The idea to deprecate this was raised on kubevirt-dev in
January 2020 and haven't got any pushback from users. Therefore, to make sure we
ship only tested code we trust, we remove support for Genie.

If a need arises in the future to support this binding, we should ideally add a
generic support for the Multu-network De-Facto standard defined by network
plumbing working group.

Signed-off-by: Petr Horacek <phoracek@redhat.com>