Open and close /dev/net/tun on virt-handler startup
Trying to access /dev/net/tun on a host which doesn't have the tun module loaded triggers a modprobe.
libvirtd in virt-launcher uses /dev/net/tun, which will trigger a modprobe if the module is not already loaded.
That won't work if/when virt-launcher runs under the SELinux type container_t.

In virt-handler, which runs as spc_t, access (open/close) /dev/net/tun and (potentially) trigger the modprobe.
That way the tun module will be loaded for sure when (libvirtd in) virt-launcher tries to access /dev/net/tun.

Signed-off-by: Jed Lejosne <jed@redhat.com>