docs: explain bridge bind mechanism
Explaining the bridge bind mechanism is a pre-requirement to
explain the masquerade bind mechanism, since plenty instructions
are common between them.

This commit explains how the pod interface configuration is
carried over to the VM interface when pod IPAM is requested, and
also how the VM is connected to the host interface (indirectly) via
an in-pod bridge created by the bind mechanism when IPAM is not
requested.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>