travis-ci: Fail on first error in script
The CI is passing several scripts of tests where each may fail.
When a script line fails, travis-ci is not existing but continues with the
next one.

This behavior makes the error hard to detect, wastes time and
resources.
Therefore, the scripts are wrapped with `set {-,+}e` for early exit on
failure.

Signed-off-by: Edward Haas <edwardh@redhat.com>