Minimize the scope of KUBEVIRT_POSTINBOUND chain
Currently, in case the vm has a list of supported ports, the source
address of the incoming traffic to the vm is substituted by the address of
k6t-eth0 bridge (relevant for both ipv4 and ipv6).

The only reason for this behaviour I could find is for connecting to the vm
using localhost address as the destination (from an app inside the
virt-launcher pod).
In this case the source address would probably be also localhost.
The vm will try to reply to local host but of course won't succeed since
the reply will stay inside the vm and won't reach the default route.
So in this specific case, the substitution of the source address
(localhost) to the bridge address is indeed needed.

However, couldn't find any other use case for this substitution.
More over, the behaviour may be harmful since the vm thinks all its
clients have the same ip address.

The commit minimizes the rules to do the substitution only in case the
source address is localhost.


Signed-off-by: Alona Kaplan <alkaplan@redhat.com>