network: report IPs list for pod network
Patch d5c5a81 introduced list of IPs to VMI interface status. This list
should deprecate the single IP report. IP report will be kept and always
report the first IP of the list. We used this entry to report IPs found
by guest agent. With this PR we enable it for pod network too, to keep
it consistent.

Example of the inconsistency:

  interfaces:
  - interfaceName: eth0
    ipAddress: 10.244.196.180
    mac: 02:00:00:b4:21:69
    name: default
  - interfaceName: eth1
    ipAddress: 192.168.66.36/24
    ipAddresses:
    - 192.168.66.36/24
    - fe80::454c:f566:baad:9107/64
    mac: 02:fb:9e:00:00:06
    name: secondary

Signed-off-by: Petr Horacek <phoracek@redhat.com>