Collect containers logs to debug CI failures
In order to track transient CI failures that are related to the infra
containers, added log collection from those containers at the end of the
suite. The logs are written to the console and not to a file because in
some cases, the artifacts collection doesn't even get to run.

Signed-off-by: Daniel Belenky <dbelenky@redhat.com>