Fix flaky test 1625 by removing outdated condition
This test, uses WaitForSuccessfulVMIStart.
The WaitForSuccessfulVMIStart helper
waits for a VMI to reach running phase.

Therefore, there is no need for another condition to wait
for the VMI to reach running.
The VMI condition manager status check is therefore removed.

The eventually clause in which the condition was examined is
also incorrect, and once it was checked, further retry on it did
not refreshed the data and just re-examined the same value.

Signed-off-by: Or Shoval <oshoval@redhat.com>