tests, vmi: Fix IPv6 configuration through the console
configureIPv6OnVMI didn't actually update the /etc/resolv.conf

Updating DNS for cirros was added as nice to have only.
The tests that actually need it, use fedora VM, which use
cloud-init for this purpose.
Hence its safe and better to remove it.
Motivation and discovery was due to flaky test 3184.

Reasons that resolv.conf wasnt updated:

 Missing sudo su for cirros.
 Using of backtick malformatted the command [1]

Adjust the expected characters when examining the echo $? response
such that it will catch the actual retcode and not the a character
which is echoed back (i.e is part of the sent command itself).

[1] https://golang.org/ref/spec#String_literals

Signed-off-by: Or Shoval <oshoval@redhat.com>