Makefile: Add timestamps to make targets
Add timestamps to the console output in order to detect where operations
time was spent, helping with debugging timeout or slow run issues.

The make targets have been wrapped with a customized shell that adds the
time stamp on each outputted line (stdout and stderr).

The customized shell includes a pipefail option ensuring that the exit status
of the pipe that passes stdout/stderr to the timestamps function is propagated
instead of being hidden behind the timestamps function that can never fail.

Signed-off-by: Daniel Belenky <dbelenky@redhat.com>