label tests of connectivity over masquerade
When KubeVirt is build on an base image that uses nftables but is
deployed on a cluster where some of the nodes have older iptables,
masquerade binding does not work (it can't call to older iptables on
hosts from newer nftables).

With this patch, we label tests that exercise connectivity over
masquerade binding so they can be skipped on environments where nftables
and iptables are mixed. For these environments, only the bridge binding
is relevant.

These tests can be skipped by passing an environment variable to the
test run:

FUNC_TEST_ARGS="--ginkgo.skip=\[label:masquerade_binding_connectivity\]" make functest

Signed-off-by: Petr Horacek <phoracek@redhat.com>