vendor: "Unreplace" prometheus-operator, update packages
vim-go was not working correctly for me, and in particular was
reporting the following error:

  vim-go: run 'go mod vendor' to sync, or use -mod=mod
          or -mod=readonly to ignore the vendor directory

When I tried running 'go mod vendor' as instructed, I got another
error:

  go: github.com/coreos/prometheus-operator@v0.0.0-20181206002233-dd6f23e7207c:
      invalid pseudo-version: does not match version-control
      timestamp (2019-02-02T02:27:40Z)

which looks like a reasonable complaint. Looking at go.mod, we

  require (
    github.com/coreos/prometheus-operator v0.35.0
  )

but at the same time we also

  replace (
    github.com/coreos/prometheus-operator =>
    github.com/coreos/prometheus-operator v0.0.0-20181206002233-dd6f23e7207
  )

The latter seems unnecessary given that we have the package vendored,
and is also the line causing the 'go mod vendor' error. Removing it
and then running

  $ make deps-update && make generate

results in this patch, which doesn't seem completely broken to my
admittedly untrained eyes and also makes vim-go happy.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>